My_Chatbot/pyproject.toml
[project]
name = "mychatbot1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
    "wsproto>=1.2.0",
]
